Main add election

See merge request raft3/kvdb-raft!7

+++ src/main/java/com/example/kvdbraft/dto/RequestVoteDTO.java
@@ -1,13 +1,16 @@
 package com.example.kvdbraft.dto;
 
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 import java.io.Serializable;
 
 @Data
-public class RequestVoteDTO implements Serializable {
+public class RequestVoteDTO implements Serializable{
     private Long term;
     private String candidateId;
-    private Long lastLogIndex;
+    private Integer lastLogIndex;
     private Long lastLogTerm;
 }
